Misc fixes (#84)
* Fixed guide formatting

* Fixed guide formatting

* Sharding bot
Updated settings
Updated tasklist
Fixed settings bug(s)

* Fixed tasks not updaing
misc fixes

* Fixed ~available not showing challenges

* Added icons
Fixed help

* Big misc fixes
characters
API fixes
fixed cancel/start not updating correctly
fixed icon names
added profile timelapse commands
fixed timezone command
added dialogue for reverting
fixed weekly report when no tasks
Fixed XPHelper calculations

* Fixed starting/skipping/reverting skills updating incorrectly
Fixed icon names

* Added some icons

* Many fixes

* Revert "Added some icons"

This reverts commit 3b18973bcb58d8db6c03b8df28249ce242e34324.

* Fixed incorrect timezone in ~tasks

* Fixed date shit

* Fixed task list showing incorrectly for non-daily tasks

* Fixed items

* Fixed tasklist length set to zero for testing

* Fixed many things

* Fixed unknown message error

* Dms enabled

* Fixed challenges showing icorrectly

* Fixed revert warning dialogue and unlocked

* Fixed page

* Fixed listpage

* Fixed timezone, fixed menu error

* fixed levelup
fixed base help

* Fixed levelup

* Fixed error thrown

* Fixed so many things

* FIXED THE BUG

* fixed unlocked null message not working

* icons

* Fixed icons

* Fixed icons

* Fixed icons

* Fixed goal being over 100 characters

* Delete Function

* Fixed lots of things

* Fixed lots of things

* Fixed SVG library error

* Fixed item description broken when link is null

* rename "reset" to "delete"
Add bot admins
Remove unnecesary join event and intent
renamed stats to bot
loading slash commands correctly
improve ~Set readability

* Misc fixes 2: The slashifying

* Fixed weekly report
Removed slash command stuff

* Revert "Misc fixes 2: The slashifying"

This reverts commit fc58df529bc59e30dc89f63ed7a87cfa2f33ae2c.

* Fixed weekly review and task list

* Fixed task list incorrectly showing goal as undefined
fixed weekly review incorrectly calculating completion percentages

* Fixed error with tasks embed field length being above 1024, split tasks into multiple fields

* Fixed timezone not updating tasks after updating tasklist

* Fixed tasklist to remove tasks when finished.
Fixed weekly report
Fixed N/A intervals

* Fixed task bug showing incorrect number of completions for the current interval

* Fixed task showing incorrect number of completions for the current interval

* Fixed "channel" undefined
Fixed tasks not showing their length of time remaining

* Fixed rendering issue with time

* Fixed interaction failing for slow update

* fix update journaling prompt when day changes

+++ objects/setting.js
@@ -51,14 +51,17 @@ class Setting {
         "4) press Done```");
       setTimeout(() => errorMsg.delete(),10000);
       onFailure();
-      initMessage.delete();
+      if (initMessage) {
+        initMessage.delete();
+      }
+      return;
     }
 
     const collector = initMessage.channel.createMessageComponentCollector({time: 120000});
     collector.on("collect", async i => {
+      await i.deferUpdate();
       callback(initMessage);
       collector.stop();
-      await i.deferUpdate();
     });
   }
   /**
